t4c/pyproject.toml
# SPDX-FileCopyrightText: Copyright DB InfraGO AG and contributors
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]

name = "t4c-cli"
requires-python = ">=3.11, <3.12"
license = { text = "Apache-2.0" }
authors = [{ name = "DB InfraGO AG" }]
keywords = []
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "License :: OSI Approved :: Apache Software License",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
]
dependencies = ["PyYAML", "click", "lxml"]

[project.urls]
Homepage = "https://github.com/DSD-DBS/capella-dockerimages"

[project.optional-dependencies]
dev = [
  "black",
  "isort",
  "mypy",
  "pylint",
  "pytest",
  "types-PyYAML",
  "types-lxml",
]

[project.scripts]
backup = "t4c_cli.backup:backup"
exporter = "t4c_cli.export:export"

[tool.black]
line-length = 79
target-version = ["py311"]

[tool.coverage.run]
branch = true
command_line = "-m pytest"
source = ["t4c_cli"]

[tool.coverage.report]
exclude_also = [
  'if t\.TYPE_CHECKING:',
  'class .*\bt\.Protocol\):',
  '@abc\.abstractmethod',
  '@t\.overload',
]
skip_covered = true

[tool.docformatter]
wrap-descriptions = 72
wrap-summaries = 79

[tool.isort]
profile = 'black'
line_length = 79

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_unreachable = true
python_version = "3.11"

[[tool.mypy.overrides]]
# Untyped third party libraries
module = [
  # ...
]
ignore_missing_imports = true

[tool.pydocstyle]
convention = "numpy"
add-select = [
  "D212", # Multi-line docstring summary should start at the first line
  "D402", # First line should not be the function’s “signature”
  "D417", # Missing argument descriptions in the docstring
]
add-ignore = [
  "D100", # Missing docstring in public module
  "D101", # Missing docstring in public class
  "D103", # Missing docstring in public function
  "D104", # Missing docstring in public package
  "D201", # No blank lines allowed before function docstring  # auto-formatting
  "D202", # No blank lines allowed after function docstring  # auto-formatting
  "D203", # 1 blank line required before class docstring  # auto-formatting
  "D204", # 1 blank line required after class docstring  # auto-formatting
  "D211", # No blank lines allowed before class docstring  # auto-formatting
  "D213", # Multi-line docstring summary should start at the second line
]

[tool.pylint.format]
ignore-long-lines = '^\s*(?:(?:__ |\.\. __: )?https?://[^ ]+$|def test_.*|[A-Za-z0-9_\.]+(?: ?:)?$)'

[tool.pylint.master]
init-import = "yes"
load-plugins = ["pylint.extensions.mccabe", "pylint.extensions.bad_builtin"]
max-complexity = 14
max-line-length = 79
extension-pkg-allow-list = ["lxml.builder", "lxml.etree"]


[tool.pylint.messages_control]
disable = [
  "broad-except",
  "global-statement",
  "import-outside-toplevel",
  "missing-class-docstring",
  "missing-function-docstring",
  "missing-module-docstring",
  "no-else-break",
  "no-else-continue",
  "no-else-raise",
  "no-else-return",
  "protected-access",
  "redefined-builtin",
  "too-few-public-methods",
  "too-many-ancestors",
  "too-many-arguments",
  "too-many-boolean-expressions",
  "too-many-branches",
  "too-many-instance-attributes",
  "too-many-lines",
  "too-many-locals",
  "too-many-public-methods",
  "too-many-return-statements",
  "too-many-statements",

  # Auto-formatting
  "bad-indentation",
  "inconsistent-quotes",
  "line-too-long",
  "missing-final-newline",
  "mixed-line-endings",
  "multiple-imports",
  "multiple-statements",
  "trailing-newlines",
  "trailing-whitespace",
  "unexpected-line-ending-format",
  "ungrouped-imports",
  "wrong-import-order",
  "wrong-import-position",

  # Handled by mypy
  "arguments-differ",
  "assignment-from-no-return",
  "import-error",
  "missing-kwoa",
  "no-member",
  "no-value-for-parameter",
  "redundant-keyword-arg",
  "signature-differs",
  "syntax-error",
  "too-many-function-args",
  "unbalanced-tuple-unpacking",
  "undefined-variable",
  "unexpected-keyword-arg",
]
enable = [
  "c-extension-no-member",
  "deprecated-pragma",
  "use-symbolic-message-instead",
  "useless-suppression",
]


[tool.pytest.ini_options]
addopts = """
  --import-mode=importlib
  --strict-config
  --strict-markers
"""
testpaths = ["tests"]
xfail_strict = true

[tool.setuptools]
platforms = ["any"]
zip-safe = false

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.setuptools.packages.find]
include = ["t4c_cli", "t4c_cli.*"]
